ansible: refactor builders with inheritance (step2)
